Move Commands to cmdguard utils package
