+++ pkg/csi/plugins/nodeserver.go
@@ -30,8 +30,8 @@ import (
 	"github.com/fluid-cloudnative/fluid/pkg/common"
 	"github.com/fluid-cloudnative/fluid/pkg/ddc/base"
 	"github.com/fluid-cloudnative/fluid/pkg/utils"
+	"github.com/fluid-cloudnative/fluid/pkg/utils/cmdguard"
 	"github.com/fluid-cloudnative/fluid/pkg/utils/dataset/volume"
-	securityutils "github.com/fluid-cloudnative/fluid/pkg/utils/security"
 	"github.com/pkg/errors"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -162,7 +162,7 @@ func (ns *nodeServer) NodePublishVolume(ctx context.Context, req *csi.NodePublis
 	} else {
 		args = append(args, mountPath, targetPath)
 	}
-	command, err := securityutils.Command("mount", args...)
+	command, err := cmdguard.Command("mount", args...)
 	if err != nil {
 		return nil, status.Error(codes.InvalidArgument, err.Error())
 	}
@@ -512,7 +512,7 @@ func checkMountInUse(volumeName string) (bool, error) {
 
 	// TODO: refer to https://github.com/kubernetes-sigs/alibaba-cloud-csi-driver/blob/4fcb743220371de82d556ab0b67b08440b04a218/pkg/oss/utils.go#L72
 	// for a better implementation
-	command, err := securityutils.Command("/usr/local/bin/check_bind_mounts.sh", volumeName)
+	command, err := cmdguard.Command("/usr/local/bin/check_bind_mounts.sh", volumeName)
 	if err != nil {
 		return inUse, err
 	}
