+++ pkg/utils/kubectl/configmap.go
@@ -21,7 +21,7 @@ import (
 	"os/exec"
 	"strings"
 
-	"github.com/fluid-cloudnative/fluid/pkg/utils/security"
+	"github.com/fluid-cloudnative/fluid/pkg/utils/cmdguard"
 	"github.com/go-logr/logr"
 	ctrl "sigs.k8s.io/controller-runtime"
 )
@@ -69,7 +69,7 @@ func kubectl(args []string) ([]byte, error) {
 
 	// return syscall.Exec(cmd, args, env)
 	// 2. execute the command
-	cmd, err := security.Command(binary, args...)
+	cmd, err := cmdguard.Command(binary, args...)
 	if err != nil {
 		return nil, err
 	}
